[NF] Load pyleecan and SciDataTool object in DataKeeper result
