+++ pyleecan/Classes/DataKeeper.py
@@ -22,6 +22,11 @@ try:
 except ImportError as error:
     as_dict = error
 
+try:
+    from ..Methods.Simulation.DataKeeper._set_result import _set_result
+except ImportError as error:
+    _set_result = error
+
 
 from ntpath import basename
 from os.path import isfile
@@ -36,6 +41,7 @@ class DataKeeper(FrozenClass):
 
     VERSION = 1
 
+    # Check ImportError to remove unnecessary dependencies in unused method
     # cf Methods.Simulation.DataKeeper.as_dict
     if isinstance(as_dict, ImportError):
         as_dict = property(
@@ -45,6 +51,17 @@ class DataKeeper(FrozenClass):
         )
     else:
         as_dict = as_dict
+    # cf Methods.Simulation.DataKeeper._set_result
+    if isinstance(_set_result, ImportError):
+        _set_result = property(
+            fget=lambda x: raise_(
+                ImportError(
+                    "Can't use DataKeeper method _set_result: " + str(_set_result)
+                )
+            )
+        )
+    else:
+        _set_result = _set_result
     # save and copy methods are available in all object
     save = save
     copy = copy
@@ -354,13 +371,6 @@ class DataKeeper(FrozenClass):
     def _get_result(self):
         """getter of result"""
         return self._result
-
-    def _set_result(self, value):
-        """setter of result"""
-        if type(value) is int and value == -1:
-            value = list()
-        check_var("result", value, "list")
-        self._result = value
 
     result = property(
         fget=_get_result,
