+++ pyleecan/Generator/ClassesRef/Simulation/DataKeeper.csv
@@ -1,6 +1,6 @@
 Variable name,Unit,Description (EN),Size,Type,Default value,Minimum value,Maximum value,,Package,Inherit,Methods,Constant Name,Constant Value,Class description
 name,-,Data name,0,str,,,,,Simulation,,as_dict,VERSION,1,Class for defining data to keep on a multi-simulation
-symbol,-,Data symbol,,str,,,,,,,,,,
+symbol,-,Data symbol,,str,,,,,,,_set_result,,,
 unit,-,Data unit,,str,,,,,,,,,,
 keeper,-,Function that takes an Output in argument and return a value,,function,None,,,,,,,,,
 error_keeper,-,"Function that takes a Simulation in argument and returns a value, this attribute enables to handle errors and to put NaN values in the result matrices",,function,None,,,,,,,,,
